requirements.txt
beautifulsoup4
lxml
requests
pyyaml